# docs/api-swagger/swagger.yaml
definitions:
...
  handlers.CreateScheduleRequest:
    properties:
      enabled:
        type: boolean
      event_topic:
        type: string
      name:
        type: string
      params:
        additionalProperties: true
        type: object
      runner_type:
        type: string
      schedule:
        description: cron expression
        type: string
      target:
        type: string
      trigger_type:
        description: cron, event, watch, manual
        type: string
      watch_path:
        type: string
      workflow_kind:
        description: '"module" or "flow"'
        type: string
      workflow_name:
        type: string
      workspace:
        type: string
    type: object
...
  handlers.UpdateScheduleRequest:
    properties:
      enabled:
        type: boolean
      name:
        type: string
      params:
        additionalProperties: true
        type: object
      schedule:
        type: string
      target:
        type: string
      workspace:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /golish/api/schedules:
    get:
      description: Get a paginated list of all scheduled workflows
      parameters:
      - default: 0
        description: Number of records to skip
        in: query
        name: offset
        type: integer
      - default: 20
        description: Maximum number of records to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of schedules
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all schedules
      tags:
      - Schedules
    post:
      consumes:
      - application/json
      description: Create a scheduled workflow execution with cron expression
      parameters:
      - description: Schedule configuration
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Schedule created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new schedule
      tags:
      - Schedules
  /golish/api/schedules/{id}:
    delete:
      description: Delete a schedule by ID
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule deleted
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Schedule not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a schedule
      tags:
      - Schedules
    get:
      description: Get details of a specific schedule by ID
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Schedule not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get schedule details
      tags:
      - Schedules
    put:
      consumes:
      - application/json
      description: Update an existing schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      - description: Schedule update data
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Schedule updated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Schedule not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a schedule
      tags:
      - Schedules
  /golish/api/schedules/{id}/disable:
    post:
      description: Disable an enabled schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule disabled
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Disable a schedule
      tags:
      - Schedules
  /golish/api/schedules/{id}/enable:
    post:
      description: Enable a disabled schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule enabled
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Enable a schedule
      tags:
      - Schedules
  /golish/api/schedules/{id}/trigger:
    post:
      description: Manually trigger a scheduled workflow execution
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Schedule triggered
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Trigger a schedule
      tags:
      - Schedules
...
swagger: "2.0"